+++ app/scripts/controllers/index.js
@@ -40,3 +40,4 @@ export * from './visualizerModal'
 export * from './channel-tabs/basicInfo'
 export * from './channel-tabs/requestMatching'
 export * from './channel-tabs/userAccess'
+export * from './channel-tabs/routes'
